+++ packages/verifiable-credentials/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Support easy creation of a Verida compatible credential record
 - Use new @verida/helpers package
 - Export interfaces
+- Support defining icon of Verida compatible credential record
 
 2022-02-09 (v2.0.6)
 -------------------